Rename Features to Plugins
see: https://github.com/prooph/event-store/issues/75
