+++ tests/Plugin/PluginManagerTest.php
@@ -9,17 +9,16 @@
  * Date: 21.04.14 - 00:16
  */
 
-namespace Prooph\EventStoreTest\Feature;
+namespace Prooph\EventStoreTest\Plugin;
 
 use Prooph\Common\Event\ProophActionEventEmitter;
 use Prooph\EventStore\Adapter\InMemoryAdapter;
 use Prooph\EventStore\Aggregate\AggregateRepository;
 use Prooph\EventStore\Aggregate\AggregateType;
 use Prooph\EventStore\Aggregate\DefaultAggregateTranslator;
-use Prooph\EventStore\Configuration\Configuration;
 use Prooph\EventStore\EventStore;
 use Prooph\EventStore\Stream\AggregateStreamStrategy;
-use Prooph\EventStoreTest\Mock\EventLoggerFeature;
+use Prooph\EventStoreTest\Mock\EventLoggerPlugin;
 use Prooph\EventStoreTest\Mock\User;
 use Prooph\EventStoreTest\TestCase;
 use Zend\ServiceManager\Config;
@@ -28,25 +27,25 @@ use Zend\ServiceManager\ServiceManager;
 /**
  * Class FeatureManagerTest
  *
- * @package Prooph\EventStoreTest\Feature
+ * @package Prooph\EventStoreTest\Plugin
  * @author Alexander Miertsch <kontakt@codeliner.ws>
  */
-class FeatureManagerTest extends TestCase
+class PluginManagerTest extends TestCase
 {
     /**
      * @test
      */
-    public function an_invokable_feature_is_loaded_by_feature_manager_and_attached_to_event_store_by_configuration()
+    public function an_invokable_plugin_is_loaded_by_plugin_manager_and_attached_to_event_store_by_configuration()
     {
-        $featureManager = new ServiceManager(new Config([
+        $pluginManager = new ServiceManager(new Config([
             "invokables" => [
-                "eventlogger" => EventLoggerFeature::class,
+                "eventlogger" => EventLoggerPlugin::class,
             ]
         ]));
 
         $eventStore = new EventStore(new InMemoryAdapter(), new ProophActionEventEmitter());
 
-        $logger = $featureManager->get('eventlogger');
+        $logger = $pluginManager->get('eventlogger');
         $logger->setUp($eventStore);
 
         $repository = new AggregateRepository(
@@ -64,7 +63,7 @@ class FeatureManagerTest extends TestCase
 
         $eventStore->commit();
 
-        $loggedStreamEvents = $featureManager->get("eventlogger")->getLoggedStreamEvents();
+        $loggedStreamEvents = $pluginManager->get("eventlogger")->getLoggedStreamEvents();
 
         $this->assertEquals(1, count($loggedStreamEvents));
     }
